feat: cadastrar candidato

+++ src/main/java/com/school/jobby/repository/CadastroRepository.java
@@ -3,5 +3,5 @@ package com.school.jobby.repository;
 import com.school.jobby.models.Cadastro;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CadastroRepository extends JpaRepository<Cadastro, Long> {
+public interface CadastroRepository extends JpaRepository<Cadastro, Integer> {
 }
